Decouple code for luc and nebraska-agent
Now building both as a new binary
