+++ .gitignore
@@ -1,1 +1,2 @@
 luc
+nebraska-agent
