+++ agent/main.go
@@ -1,0 +1,7 @@
+package main
+
+import agentcli "github.com/kinvolk/lokomotive-update-controller/cli/agent"
+
+func main() {
+	agentcli.Execute()
+}
