+++ cli/updater/cli.go
@@ -1,0 +1,1 @@
+package updatercli
