+++ cli/updater/util.go
@@ -1,0 +1,1 @@
+package updatercli
