--- cli/util.go
@@ -1,1 +1,0 @@
-package cli
