--- main.go
@@ -1,7 +1,0 @@
-package main
-
-import "github.com/kinvolk/lokomotive-update-controller/cli"
-
-func main() {
-	cli.Execute()
-}
