+++ pkg/updater/updater.go
@@ -26,10 +26,6 @@ import (
 	"github.com/kinvolk/fluxlib/lib"
 	helmrelease "github.com/kinvolk/fluxlib/lib/helm-release"
 	sourcecontroller "github.com/kinvolk/fluxlib/lib/source-controller"
-
-	dockertypes "github.com/docker/docker/api/types"
-	dockercontainer "github.com/docker/docker/api/types/container"
-	dockerclient "github.com/docker/docker/client"
 )
 
 const (
@@ -133,42 +129,6 @@ func Reconcile(cfg *Config) error {
 
 		return false, nil
 	})
-
-	return nil
-}
-
-func ReconcileContainer(cfg *Config) error {
-	kubeconfig, err := ioutil.ReadFile(cfg.Kubeconfig)
-	if err != nil && !os.IsNotExist(err) {
-		return fmt.Errorf("reading kubeconfig: %w", err)
-	}
-
-	if err := cfg.getClusterID(kubeconfig); err != nil {
-		return fmt.Errorf("retrieving cluster id: %w", err)
-	}
-	cfg.currentVersion = defaultVersion
-
-	if err := cfg.setupNebraskaClient(); err != nil {
-		return fmt.Errorf("setting up nebraska client: %w", err)
-	}
-
-	log.Debug("initialization complete")
-
-	if cfg.Docker {
-		_ = wait.PollInfinite(time.Duration(cfg.Interval)*time.Minute, func() (done bool, err error) {
-			log.Debug("reconciling infinitely!")
-
-			if err := cfg.reconcileContainer(); err != nil {
-				log.Error(err)
-			}
-
-			return false, nil
-		})
-	} else {
-		if err := cfg.systemdService(); err != nil {
-			log.Error(err)
-		}
-	}
 
 	return nil
 }
@@ -387,166 +347,6 @@ func (cfg *Config) reconcile() error {
 
 		return err
 	}
-
-	// Update the current version to the new one.
-	cfg.currentVersion = version
-
-	_ = cfg.nbsClient.ReportProgress(ctx, updater.ProgressInstallationFinished)
-	_ = cfg.nbsClient.ReportProgress(ctx, updater.ProgressUpdateComplete)
-
-	cfg.nbsClient.SetInstanceVersion(info.Version)
-
-	return nil
-}
-
-func (cfg *Config) reconcileContainer() error {
-	ctx := context.TODO()
-
-	// Let us check if there is an update.
-	info, err := cfg.nbsClient.CheckForUpdates(ctx)
-	if err != nil {
-		return fmt.Errorf("checking for updates: %w", err)
-	}
-
-	// There is no update hence return.
-	if !info.HasUpdate {
-		log.Info("no update available")
-
-		// Print the response just in case.
-		log.Debugf("got this response: %#v", info.OmahaResponse().Apps[0])
-
-		return nil
-	}
-
-	_ = cfg.nbsClient.ReportProgress(ctx, updater.ProgressDownloadStarted)
-
-	// There is a new update.
-	version := info.Version
-	link := info.URL()
-	name := info.Package().Name
-	containerStopDuration := time.Minute
-
-	log.Debugf("update available: %s", version, link, name)
-
-	cli, err := dockerclient.NewClientWithOpts(dockerclient.FromEnv)
-	if err != nil {
-		_ = cfg.nbsClient.ReportProgress(ctx, updater.ProgressError)
-		return fmt.Errorf("setting up docker client: %w", err)
-	}
-
-	containers, err := cli.ContainerList(context.Background(), dockertypes.ContainerListOptions{})
-	if err != nil {
-		_ = cfg.nbsClient.ReportProgress(ctx, updater.ProgressError)
-		return fmt.Errorf("listing containers: %w", err)
-	}
-
-	for _, container := range containers {
-
-		imageName := strings.Split(container.Names[0], "/")
-
-		if imageName[1] == name {
-
-			log.Debug("Stopping and restaring the container")
-
-			err := cli.ContainerStop(context.Background(), container.ID[:10], &containerStopDuration)
-			if err != nil {
-				_ = cfg.nbsClient.ReportProgress(ctx, updater.ProgressError)
-				return fmt.Errorf("stopping container: %w", err)
-			}
-
-			err = cli.ContainerRemove(context.Background(), container.ID[:10], dockertypes.ContainerRemoveOptions{})
-			if err != nil {
-				_ = cfg.nbsClient.ReportProgress(ctx, updater.ProgressError)
-				return fmt.Errorf("removing container: %w", err)
-			}
-
-			// Now run a container
-			resp, err := cli.ContainerCreate(ctx, &dockercontainer.Config{
-				Image: name + ":" + version,
-			}, nil, nil, nil, name)
-			if err != nil {
-				_ = cfg.nbsClient.ReportProgress(ctx, updater.ProgressError)
-				return fmt.Errorf("creating container: %w", err)
-			}
-
-			if err := cli.ContainerStart(ctx, resp.ID, dockertypes.ContainerStartOptions{}); err != nil {
-				_ = cfg.nbsClient.ReportProgress(ctx, updater.ProgressError)
-				return fmt.Errorf("starting container: %w", err)
-			}
-		} else {
-			continue
-		}
-	}
-
-	// Update the current version to the new one.
-	cfg.currentVersion = version
-
-	_ = cfg.nbsClient.ReportProgress(ctx, updater.ProgressInstallationFinished)
-	_ = cfg.nbsClient.ReportProgress(ctx, updater.ProgressUpdateComplete)
-
-	cfg.nbsClient.SetInstanceVersion(info.Version)
-
-	return nil
-}
-
-func (cfg *Config) systemdService() error {
-	ctx := context.TODO()
-
-	// Let us check if there is an update.
-	info, err := cfg.nbsClient.CheckForUpdates(ctx)
-	if err != nil {
-		return fmt.Errorf("checking for updates: %w", err)
-	}
-
-	// There is no update hence return.
-	if !info.HasUpdate {
-		// Print the response just in case.
-		log.Debugf("got this response: %#v", info.OmahaResponse().Apps[0])
-
-		return nil
-	}
-
-	_ = cfg.nbsClient.ReportProgress(ctx, updater.ProgressDownloadStarted)
-
-	// There is a new update.
-	version := info.Version
-	// link := info.URL()
-	// name := info.Package().Name
-
-	// Read the contents of env file
-	envFile, err := ioutil.ReadFile(cfg.EnvPath)
-	if err != nil && !os.IsNotExist(err) {
-		_ = cfg.nbsClient.ReportProgress(ctx, updater.ProgressError)
-		return fmt.Errorf("reading env file: %w", err)
-	}
-
-	envData := strings.Split(string(envFile), "\n")
-	versionEnv := strings.Split(envData[1], "=")
-
-	if versionEnv[1] == version {
-		log.Info("no update available")
-		return nil
-	}
-
-	log.Debugf("update available: %s", version)
-
-	// update env file.
-	indexVersion := int64(strings.Index(string(envFile), "VERSION=")) + 8
-	file, err := os.OpenFile(cfg.EnvPath, os.O_RDWR, 0644)
-
-	if err != nil {
-		_ = cfg.nbsClient.ReportProgress(ctx, updater.ProgressError)
-		return fmt.Errorf("failed opening file: %w", err)
-	}
-
-	defer file.Close()
-
-	_, err = file.WriteAt([]byte(version), indexVersion)
-	if err != nil {
-		return fmt.Errorf("failed writing to file: %w", err)
-	}
-
-	log.Info("Updated env file to latest version.")
 
 	// Update the current version to the new one.
 	cfg.currentVersion = version
