+++ updater/main.go
@@ -1,0 +1,9 @@
+package main
+
+import (
+	updatercli "github.com/kinvolk/lokomotive-update-controller/cli/updater"
+)
+
+func main() {
+	updatercli.Execute()
+}
